# openapi/advantage-api.yaml
openapi: 3.1.0
info:
  title: Advantage API
  version: 0.1.0
  summary: The route handlers under src/app/api, as the dashboard calls them.
  description: |
    Every path here is a Next.js route handler in `src/app/api/`. The handler
    code is the source of truth: each status code below is one the handler, or
    the module it delegates to, returns as a literal.

    **Authentication.** Signed-in routes read the Supabase session from its
    cookie (`cookieAuth`) through `createClient()`; there is no bearer token.
    `/api/programs/search` is the only public browser route in this file.

    **Machine callers.** The rule above is about browser routes. The vendor
    and Stripe webhooks and the cleanup cron carry no session: they
    authenticate with `splitstepHmac`, `stripeSignature` and `cronBearer`.
    Two further operations take no credential at all — the vendor webhook's
    GET liveness check and the retired `/api/create-checkout-session`, which
    answers 410 to everyone. The admin upload routes use the session and
    additionally require `users.is_admin`.

    **Same origin.** Every mutation under `/api/matches/{matchId}/video`
    also requires a request from a page this deployment served
    (`checkSameOrigin`): a cross-site `Sec-Fetch-Site`, a missing `Origin`, or
    an `Origin` that is neither the request's own nor the configured site URL
    is refused with `403 cross_origin` before anything else runs. Those
...
    4096 bytes (`413 request_too_large`).

...

...

...
servers:
...
security:
  - cookieAuth: []
...
paths:
  /api/matches/{matchId}:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    get:
      tags: [Matches]
      operationId: getMatch
      summary: Read a match for the Edit Match dialog
      description: >-
        The caller's own match (`created_by` is the session user), with the
        latest analysis status, the event line it is filed under, and whether
        this viewer may attach it to a line. A malformed id, a missing match
        and someone else's match all answer 404.
      responses:
        "200":
          description: The match and the context the dialog lays itself out from.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchDetailResponse"
              examples:
                personalMatch:
                  summary: A hand-scored personal match
                  value:
                    match:
                      id: 5b0f4c1e-8a7d-4b43-9f7e-1c2d3e4f5a6b
                      tournament_name: Spring Open
                      round: R16
                      date: "2026-09-12T12:00:00"
                      match_type: tournament
                      court_type: hard
                      player1_id: 0e9a2b7c-4d1f-4a8e-9b3c-2f1e0d9c8b7a
                      player1_name: Alex Rivera
                      player2_name: Sam Chen
                      score:
                        player1: [6, 4, 6]
                        player2: [3, 6, 2]
                        winner: player1
                      private: true
                      format:
                        best_of: 3
                        ad_scoring: true
                      duration: null
                      player_hand: right
                      player_backhand: two-handed
                      opponent_hand: left
                      opponent_backhand: one-handed
                      program_id: null
                      event_entry_id: null
                      source_provider: null
                    analysis: null
                    event: null
                    canAttach: false
                    canDetach: false
                    canEditRound: false
        "401":
          $ref: "#/components/responses/PlainUnauthorized"
        "404":
          $ref: "#/components/responses/PlainNotFound"
        "500":
          description: The match or its analysis status could not be read.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Could not load the match
    patch:
      tags: [Matches]
      operationId: updateMatch
      summary: Edit a match
      description: >-
        Partial update of the caller's own match. The rules live in
        `normalizeMatchPatch` (`src/lib/matches/patch-match.ts`): a match on a
        scheduled event line refuses the event-owned fields, an analysed match
        keeps its format fixed, and a team match's player must come from the
        program roster (the saved name is the roster's, never the client's).
        `private` is forced to `true` on every save during the beta.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchPatchRequest"
            examples:
              rename:
                summary: Correct the opponent and the score
                value:
                  player2_name: Sam Chen
                  score:
                    player1: [6, 4, 7]
                    player2: [3, 6, 6]
                    player1_tiebreaks: [null, null, 7]
                    player2_tiebreaks: [null, null, 4]
      responses:
        "200":
          description: The saved row.
          content:
            application/json:
              schema:
                type: object
                required: [match]
                properties:
                  match:
                    $ref: "#/components/schemas/Match"
              example:
                match:
                  id: 5b0f4c1e-8a7d-4b43-9f7e-1c2d3e4f5a6b
                  tournament_name: Spring Open
                  round: R16
                  date: "2026-09-12T12:00:00"
                  match_type: tournament
                  court_type: hard
                  player1_id: 0e9a2b7c-4d1f-4a8e-9b3c-2f1e0d9c8b7a
                  player1_name: Alex Rivera
                  player2_name: Sam Chen
                  score:
                    player1: [6, 4, 7]
                    player2: [3, 6, 6]
                    player1_tiebreaks: [null, null, 7]
                    player2_tiebreaks: [null, null, 4]
                    winner: player1
                  private: true
                  format:
                    best_of: 3
                  duration: null
                  player_hand: right
                  player_backhand: two-handed
                  opponent_hand: null
                  opponent_backhand: null
                  program_id: null
                  event_entry_id: null
                  source_provider: null
        "400":
          description: >-
            The body is not a JSON object, or a field is refused. A field
            refusal names the field so the dialog can mark it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldErrorResponse"
              examples:
                invalidJson:
                  summary: Body is not a JSON object
                  value:
                    error: Invalid JSON body
                eventOwned:
                  summary: An event-owned field on a scheduled match
                  value:
                    error: This match is on a scheduled event, so that comes from the event. Change it in Schedule.
                    field: date
                offRoster:
                  summary: A player who is not on the program roster
                  value:
                    error: Choose a player on this team's roster.
                    field: player1_id
                empty:
                  summary: No editable field in the body
                  value:
                    error: Nothing to save.
        "401":
          $ref: "#/components/responses/PlainUnauthorized"
        "404":
          $ref: "#/components/responses/PlainNotFound"
        "500":
          description: The match, its analysis status or the roster could not be read, or the save failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                load:
                  value:
                    error: Could not load the match
                roster:
                  value:
                    error: Could not load the roster
                save:
                  value:
                    error: Could not save the match
    delete:
      tags: [Matches]
      operationId: deleteMatch
      summary: Delete a match and its stored files
      description: >-
        Purges the match's storage first (`purgeMatchStorage`), then deletes
        the row. A match recorded or analysed through the admin console is
        refused with 409; a purge that cannot run is a retryable 503.
      responses:
        "200":
          description: The match is gone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
              example:
                ok: true
        "401":
          $ref: "#/components/responses/PlainUnauthorized"
        "404":
          $ref: "#/components/responses/PlainNotFound"
        "409":
          description: The admin console's purge guard protects this match.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Matches recorded or analyzed through the admin console cannot be deleted here.
        "500":
          description: The match could not be read, or the row delete failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                load:
                  value:
                    error: Could not load the match
                delete:
                  value:
                    error: Could not delete the match
        "503":
          description: The storage purge could not run. Nothing was deleted; retry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Match deletion is unavailable. Try again.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-pouxujkhtbvkdwbzfvka-auth-token
      description: >-
        The Supabase Auth session cookie set by `@supabase/ssr`, named
        `sb-<project-ref>-auth-token` (the production project ref is shown;
        another project uses its own). A large session is split into
        `.0`, `.1`, … chunks. `src/proxy.ts` refreshes it on every request.

    cronBearer:
      type: http
      scheme: bearer
      description: >-
        `Authorization: Bearer <CRON_SECRET>`, as Vercel Cron sends it. The
        scheme word is matched case-insensitively; the credential is compared
        as SHA-256 digests (`authorizeCronRequest()` in
        `src/lib/services/match-video/cleanup-schedule.ts`). An unset or blank
        `CRON_SECRET` refuses every request.
    splitstepHmac:
      type: apiKey
      in: header
      name: x-hmac-signature
      description: >-
        `base64(HMAC-SHA256(SPLITSTEP_WEBHOOK_SECRET, raw_body))`, computed
        over the exact bytes received (`verifyWebhookAuth()` in
        `src/lib/services/splitstep/webhook-auth.ts`). Also looked for, as a
        hedge, under `x-splitstep-signature`, `x-webhook-signature`,
        `x-signature`, `x-signature-256`, `x-hub-signature-256`, `signature`,
        `x-webhook-secret`, `x-api-key` and `authorization` (a `Bearer `
        prefix is stripped). Every one of those headers is redacted before
        the delivery is recorded.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: >-
        Stripe's webhook signature, verified by
        `stripe.webhooks.constructEvent()` against `STRIPE_WEBHOOK_SECRET`
        over the raw body.
  parameters:
    MatchId:
      name: matchId
      in: path
      required: true
      description: The match's id. A value that is not a UUID answers 404, never 400.
      schema:
        type: string
        format: uuid
      example: 5b0f4c1e-8a7d-4b43-9f7e-1c2d3e4f5a6b
...
  responses:
    PlainUnauthorized:
      description: No session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Unauthorized
    PlainNotFound:
      description: >-
        The id is not a UUID, the match does not exist, or it is not the
        caller's. The three are indistinguishable on purpose.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Not found
...
  schemas:
    ErrorResponse:
      type: object
      description: >-
        The error envelope. `error` is the sentence a client may show; `code`
        is the stable slug a client branches on (video routes only); `detail`
        is a cause slug for logs and tests — never rendered, and a client must
        not depend on it being present.
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        detail:
          type: string
    FieldErrorResponse:
      description: An `ErrorResponse` that may name the refused field.
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            field:
              type: string
              description: The body key the refusal is about.
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          const: true
...
    Match:
      type: object
      description: The `matches` columns the Edit Match dialog reads.
      required: [id]
      properties:
        id:
          type: string
          format: uuid
        tournament_name:
          type: [string, "null"]
        round:
          type: [string, "null"]
        date:
          type: [string, "null"]
          description: Stored as local noon, `YYYY-MM-DDT12:00:00`.
        match_type:
          type: [string, "null"]
        court_type:
          type: [string, "null"]
        player1_id:
          type: [string, "null"]
          description: An auth user id or a `program_players.id`; never normalise it.
        player1_name:
          type: [string, "null"]
        player2_name:
          type: [string, "null"]
        score:
          oneOf:
            - $ref: "#/components/schemas/MatchScore"
            - type: "null"
        private:
          type: boolean
        format:
          oneOf:
            - $ref: "#/components/schemas/MatchFormat"
            - type: "null"
        duration:
          type: [string, integer, "null"]
        player_hand:
          type: [string, "null"]
        player_backhand:
          type: [string, "null"]
        opponent_hand:
          type: [string, "null"]
        opponent_backhand:
          type: [string, "null"]
        program_id:
          type: [string, "null"]
          format: uuid
        event_entry_id:
          type: [string, "null"]
          format: uuid
        source_provider:
          type: [string, "null"]
          description: "`swing-vision`, `splitstep`, or null for a hand-scored match."
    AnalysisStatus:
      type: string
      enum:
        - uploading
        - queued
        - processing
        - deriving
        - completed
        - failed
        - derivation_failed
        - uploaded
    MatchEventContext:
      type: object
      required:
        [
          eventId,
          eventName,
          eventKind,
          slot,
          discipline,
          startsOn,
          endsOn,
          site,
          surface,
          format,
          takenRounds,
        ]
      properties:
        eventId:
          type: string
          format: uuid
        eventName:
          type: string
        eventKind:
          type: string
          enum: [dual, tournament]
        slot:
          type: [string, "null"]
        discipline:
          type: string
        startsOn:
          type: string
          format: date
        endsOn:
          type: string
          format: date
        site:
          type: string
        surface:
          type: [string, "null"]
        format:
          type: [object, "null"]
          properties:
            best_of:
              type: integer
            ad_scoring:
              type: [boolean, "null"]
        takenRounds:
          type: array
          items:
            type: string
          description: >-
            Round codes the line already holds without this match (its other
            matches and its outcomes). The Round menu leaves them out, and
            `set_match_round_on_line` refuses them.
    MatchDetailResponse:
      type: object
      required: [match, analysis, event, canAttach, canDetach, canEditRound]
      properties:
        match:
          $ref: "#/components/schemas/Match"
        analysis:
          description: The newest processing job's status, or null when there is none.
          oneOf:
            - type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/AnalysisStatus"
            - type: "null"
        event:
          description: The event line the match is filed under, or null.
          oneOf:
            - $ref: "#/components/schemas/MatchEventContext"
            - type: "null"
        canAttach:
          type: boolean
          description: >-
            A team match not yet on a line, in the viewed team workspace, by
            someone who may run its schedule.
        canDetach:
          type: boolean
          description: >-
            A team match that is on a line, in the viewed team workspace, by
            someone who may run its schedule.
        canEditRound:
          type: boolean
          description: >-
            `canDetach` on a tournament line, whose round is the match's own
            (a dual line's round is its slot).
    MatchPatchRequest:
      type: object
      description: >-
        Every key is optional; send only what changed. At least one editable
        key is required, or the answer is 400 "Nothing to save.". On a match
        filed under an event line, `tournament_name`, `round`, `date`,
        `match_type` and `court_type` are refused.
      minProperties: 1
      properties:
        tournament_name:
          type: [string, "null"]
        round:
          type: [string, "null"]
        date:
          type: string
          format: date
        match_type:
          type: [string, "null"]
        court_type:
          type: [string, "null"]
        player1_id:
          type: string
          description: Team matches only — a player on the program roster.
        player1_name:
          type: string
          minLength: 1
        player2_name:
          type: string
          minLength: 1
        player_hand:
          oneOf:
            - $ref: "#/components/schemas/Hand"
            - type: "null"
        opponent_hand:
          oneOf:
            - $ref: "#/components/schemas/Hand"
            - type: "null"
        player_backhand:
          oneOf:
            - $ref: "#/components/schemas/Backhand"
            - type: "null"
        opponent_backhand:
          oneOf:
            - $ref: "#/components/schemas/Backhand"
            - type: "null"
        format:
          description: Refused on an analysed or event-linked match.
          allOf:
            - $ref: "#/components/schemas/MatchFormat"
        score:
          $ref: "#/components/schemas/MatchScore"
...